Formats boundbox for WFS query

Convenience function that formats a spatial bounding box into a character string that can be passed to the WFS query. The function accepts either a bbox from st_bbox or the coordinates for the box edges (xmin, ymin, xmax, ymax).

boundbox(xmin, ymin, xmax, ymax, st_bbox = NULL)

Arguments

xmin
numeric; minimum x coordinate (longitude)

ymin
numeric; minimum y coordinate (latitude)

xmax
numeric; maximum x coordinate (longitude)

ymax
numeric; maximum y coordinate (latitude)

st_bbox
bbox; object generated using st_bbox

Value

character string formatted for WFS query and use in read_layer_sf

Examples

#### Using an sf object ####
sf_data <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
boundbox(st_bbox = sf::st_bbox(sf_data))

#> [1] "-84.3238525390625,33.8819923400879,-75.4569778442383,36.5896492004395"

#### Using coordinates ####
boundbox(xmin = 144.05, ymin = -38.44, xmax = 144.50, ymax = -38.10)

#> [1] "144.05,-38.44,144.5,-38.1"