Check Geoserver Response

VicmapR relies upon a functioning geoserver. If for whatever reason the geoserver is not functioning then the functions in this package will not work. This function will check the response of the geoserver; erroring out if the connection is down.

check_geoserver(timeout = 15, quiet = FALSE)

Arguments

timeout
numeric: the time (in seconds) to wait for the response before timing out (default is 15)

quiet
logical: whether to silently check the connection and if working, return nothing. If FALSE (default), the status message will be printed (http_status)

Value

logical, TRUE if the geoserver is working

Examples

# \donttest{
try(
check_geoserver()
)

#> [1] TRUE

# }