Convert layer names from old platform naming convention to the new platform naming convention

Convert layer names from old platform naming convention to the new platform naming convention

convert_layer_name(x)

Arguments

x

object of class vicmap_promise

Value

object of class vicmap_promise

Examples

# \donttest{
try(
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") 
)
#> You are using old layer names. We converted your layer to hy_watercourse with a CQL filter of feature_type_code='watercourse_channel_drain'. To suppress this message, update your code to use the new layer names (see VicmapR::name_conversions for more details)
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid
#> • Using collect() on this object will return 195349
#> • features and 21 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 20 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 141.761 ymin: -34.36473 xmax: 142.3621 ymax: -34.34771
#> Geodetic CRS:  GDA94
#> # A tibble: 6 × 21
#>   id                  ufi    pfi feature_type_code name  named_feature_id origin
#>   <chr>             <int>  <int> <chr>             <chr> <chr>            <chr> 
#> 1 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> 2 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> 3 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> 4 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> 5 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> 6 hy_watercourse.… 2.55e6 8.55e6 watercourse_chan… NA    NA               2     
#> # ℹ 14 more variables: construction <chr>, usage <chr>, hierarchy <chr>,
#> #   feature_quality_id <int>, create_date_pfi <dttm>, superceded_pfi <chr>,
#> #   create_date_ufi <dttm>, auth_org_code <chr>, auth_org_id <chr>,
#> #   auth_org_verified <chr>, task_id <chr>, feature_ufi <int>,
#> #   feature_create_date_ufi <dttm>, geometry <LINESTRING [°]>
# }