Layer Metadata

formatted metadata attributes of a given vicmap layer (vicmap_query(layer)). Metadata is retrieved from the Vicmap catalogue. data_citation() prints a BibTex style citation for a given record; similar to base::citation(). data_dictionary() returns a table with names, types and descriptions of the data within the selected layer (see details). get_metdata() returns a list with three elements, containing metadata, the data dictionary and the url of the metadata for the record.

data_citation(x = NULL, anzlicId = NULL)

data_dictionary(x = NULL, anzlicId = NULL)

get_metadata(x = NULL, anzlicId = NULL)

Arguments

x

Object of class vicmap_promise (likely passed from vicmap_query())

anzlicId

character: ID of data (useful if data is not available through WFS)

Value

citation, data.frame or list

Examples

# \donttest{
try(
data_citation(vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN"))
)
#>   @ELECTRONIC{HY_WATERCOURSE,
#>         author = {Department of Environment, Land, Water & Planning},
#>         title = {Watercourse Network 1:25,000 - Vicmap Hydro},
#>         year = {2020},
#>         url = {http://services.land.vic.gov.au/catalogue/metadata?anzlicId=ANZVI0803002490&publicId=guest&extractionProviderId=1},
#>         owner = {Department of Environment, Land, Water & Planning},
#>         timestamp = {2022.03.07},
#> }
# }
# \donttest{
try(
data_dictionary(vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN"))
)
#> # A tibble: 13 x 7
#>    `Column Name` `Column Name 10` Obligation Unique `Data Type` `Reference Tab~`
#>    <chr>         <chr>            <chr>      <chr>  <chr>       <chr>           
#>  1 CONSTRUCTION  CONST            O          N      VARCHAR2(1) VMREFTAB.HY_CON~
#>  2 CREATE_DATE_~ CRDATE_PFI       O          N      DATE        NA              
#>  3 CREATE_DATE_~ CRDATE_UFI       O          N      DATE        NA              
#>  4 FEATURE_QUAL~ FQID             O          N      NUMBER(9)   NA              
#>  5 FEATURE_TYPE~ FTYPE_CODE       O          N      VARCHAR2(3~ NA              
#>  6 HIERARCHY     HIERARCHY        O          N      VARCHAR2(1) VMREFTAB.HY_HIE~
#>  7 NAME          NAME             O          N      VARCHAR2(5~ NA              
#>  8 NAMED_FEATUR~ NFEAT_ID         O          N      NUMBER(9)   NA              
#>  9 ORIGIN        ORIGIN           O          N      VARCHAR2(1) VMREFTAB.HY_ORI~
#> 10 PFI           PFI              M          Y      NUMBER(9)   NA              
#> 11 SUPERCEDED_P~ SUPER_PFI        O          N      NUMBER(9)   NA              
#> 12 UFI           UFI              O          N      NUMBER(9)   NA              
#> 13 USAGE         USAGE            O          N      VARCHAR2(1) VMREFTAB.HY_USA~
#> # ... with 1 more variable: Comments <chr>
# }
# \donttest{
try(
get_metadata(vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN"))
)
#> [[1]]
#> # A tibble: 30 x 2
#>    `Metadata Name`           Description                                        
#>    <chr>                     <chr>                                              
#>  1 Resource Name             "HY_WATERCOURSE"                                   
#>  2 Title                     "Watercourse Network 1:25,000 - Vicmap Hydro"      
#>  3 Anzlic Id                 "ANZVI0803002490\r\n\t\t\t\t\t\t\r\n\t\t\t\t\t\t\t~
#>  4 Custodian                 "Department of Environment, Land, Water & Planning"
#>  5 Owner                     "Department of Environment, Land, Water & Planning"
#>  6 Jurisdiction              "Victoria"                                         
#>  7 Abstract                  "This layer is part of Vicmap Hydro and contains l~
#>  8 Search Words              "LAND Topography, AGRICULTURE Irrigation, WATER Hy~
#>  9 Purpose                   "Vicmap Hydro provides an accurate representation ~
#> 10 Geographic Extent Polygon  NA                                                
#> # ... with 20 more rows
#> 
#> [[2]]
#> # A tibble: 13 x 7
#>    `Column Name` `Column Name 10` Obligation Unique `Data Type` `Reference Tab~`
#>    <chr>         <chr>            <chr>      <chr>  <chr>       <chr>           
#>  1 CONSTRUCTION  CONST            O          N      VARCHAR2(1) VMREFTAB.HY_CON~
#>  2 CREATE_DATE_~ CRDATE_PFI       O          N      DATE        NA              
#>  3 CREATE_DATE_~ CRDATE_UFI       O          N      DATE        NA              
#>  4 FEATURE_QUAL~ FQID             O          N      NUMBER(9)   NA              
#>  5 FEATURE_TYPE~ FTYPE_CODE       O          N      VARCHAR2(3~ NA              
#>  6 HIERARCHY     HIERARCHY        O          N      VARCHAR2(1) VMREFTAB.HY_HIE~
#>  7 NAME          NAME             O          N      VARCHAR2(5~ NA              
#>  8 NAMED_FEATUR~ NFEAT_ID         O          N      NUMBER(9)   NA              
#>  9 ORIGIN        ORIGIN           O          N      VARCHAR2(1) VMREFTAB.HY_ORI~
#> 10 PFI           PFI              M          Y      NUMBER(9)   NA              
#> 11 SUPERCEDED_P~ SUPER_PFI        O          N      NUMBER(9)   NA              
#> 12 UFI           UFI              O          N      NUMBER(9)   NA              
#> 13 USAGE         USAGE            O          N      VARCHAR2(1) VMREFTAB.HY_USA~
#> # ... with 1 more variable: Comments <chr>
#> 
#> [[3]]
#> [1] "http://services.land.vic.gov.au/catalogue/metadata?anzlicId=ANZVI0803002490&publicId=guest&extractionProviderId=1"
#> 
# }