Filter Data

See dplyr::filter for details.

# S3 method for vicmap_promise
filter(.data, ...)

Arguments

.data
object of class vicmap_promise (likely passed from vicmap_query())

...
filter statements

Value

Object of class vicmap_promise, which is a 'promise' of the data that can be returned if collect() is used

Methods (by class)

vicmap_promise: filter.vicmap_promise

Examples

# \donttest{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>%
 filter(HIERARCHY == "L", PFI == 8553127)

#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#ident’.
#>  "wfsConnection#ANY" would also be valid

#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid

#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteString’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid

#> • Using collect() on this object will return 1 features
#> • and 16 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 1 feature and 15 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 143.3209 ymin: -35.06905 xmax: 143.324 ymax: -35.06721
#> Geodetic CRS:  GDA94
#> # A tibble: 1 x 16
#>   id                  PFI    UFI FEATURE_TYPE_CODE NAME  NAMED_FEATURE_ID ORIGIN
#>   <chr>             <int>  <int> <chr>             <chr> <chr>            <chr> 
#> 1 VMHYDRO_WATERC… 8553127 2.55e6 watercourse_chan… NA    NA               2     
#> # … with 9 more variables: CONSTRUCTION <chr>, USAGE <chr>, HIERARCHY <chr>,
#> #   FEATURE_QUALITY_ID <int>, CREATE_DATE_PFI <dttm>, SUPERCEDED_PFI <chr>,
#> #   CREATE_DATE_UFI <dttm>, OBJECTID <int>, geometry <LINESTRING [°]>

 # }