List Available WFS Layers

Lists layers available from the WFS geoserver. This is similar to sending the WFS request of getFeatureTypes. listLayers() returns a data.frame with the 'Name' and title of the layers available. The 'Name' is what is used within vicmap_query() while the title provides somewhat of a description/clarification about the layer.

listLayers(..., abstract = TRUE)

Arguments

...

Additional arguments passed to grep. The pattern argument can be used to search for specific layers with matching names or titles.

abstract

Whether to return a column of abstract (and metadata ID), the default is true. Switching to FALSE will provide a data.frame with only 2 columns and may be slightly faster.

Value

data.frame of 2 (abstract = FALSE) or 4 (abstract = TRUE) columns

Examples

# \donttest{
try(
listLayers(pattern = "trees", ignore.case = TRUE)
)
#>                              Name        Title
#> 1 open-data-platform:tree_density tree_density
#> 2   open-data-platform:tree_urban   tree_urban
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Abstract
#> 1 The Vicmap Vegetation Tree Extent dataset was generalised to 2m pixels and then clustering rules were applied to group the data into three density classes (Dense, Medium, Sparse). This classification was a pixel by pixel assessment where a pixel was allocated a density classification based on neighbouring pixels. The raster dataset was then converted to vector. \n\nThe process of grouping tree cover into density classes simplifies the representation of trees and reduces the complexity of the vector dataset. It is a effective way of representing tree cover. The original raw 20cm raster dataset is maintained as a separate dataset, Vicmap Vegetation Tree Extent.
#> 2                                                                                                                                                                                                                                                                                                                                                           The Vicmap Vegetation Tree Urban represents trees as points across Metropolitan Melbourne and the urban environment within four regional councils: Wangaratta, Sale, Shepparton and Ballarat. \n\nThis product is derived from machine learning of high resolution aerial photography with no post processing human intervention.
#>                             metadataID
#> 1 1af4942a-fff4-58c2-80f7-10c6054c31c4
#> 2 51eba84d-7cb2-55bc-bf96-f10cb44bc923
# }