Print a Snapshot of the Data

print() displays a cut of the data (no more than six rows) alongside the number of rows and columns that would be returned.

# S3 method for vicmap_promise
print(x, ...)

Arguments

x
object of class vicmap_promise (likely passed from vicmap_query())

...
arguments to be passed to print

Value

vicmap_promise (invisible), promise sample printed to console

Examples

# \donttest{
try(
query <- vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN")
)
try(
print(query)
)
#> • Using collect() on this object will return 191506
#> • features and 16 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 6 features and 15 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 142.7675 ymin: -35.06905 xmax: 143.324 ymax: -35.04559
#> Geodetic CRS:  GDA94
#> # A tibble: 6 × 16
#>   id                  PFI    UFI FEATURE_TYPE_CODE NAME  NAMED_FEATURE_ID ORIGIN
#>   <chr>             <int>  <int> <chr>             <chr> <chr>            <chr> 
#> 1 VMHYDRO_WATERC… 8553127 2.55e6 watercourse_chan… NA    NA               2     
#> 2 VMHYDRO_WATERC… 8553130 2.55e6 watercourse_chan… NA    NA               2     
#> 3 VMHYDRO_WATERC… 8553143 2.55e6 watercourse_chan… NA    NA               2     
#> 4 VMHYDRO_WATERC… 8553149 2.55e6 watercourse_chan… NA    NA               2     
#> 5 VMHYDRO_WATERC… 8553158 2.55e6 watercourse_chan… NA    NA               2     
#> 6 VMHYDRO_WATERC… 8553168 2.55e6 watercourse_chan… NA    NA               2     
#> # … with 9 more variables: CONSTRUCTION <chr>, USAGE <chr>, HIERARCHY <chr>,
#> #   FEATURE_QUALITY_ID <int>, CREATE_DATE_PFI <dttm>, SUPERCEDED_PFI <chr>,
#> #   CREATE_DATE_UFI <dttm>, OBJECTID <int>, geometry <LINESTRING [°]>
# }