Read Vicmap WFS layer as sf

Reads in a Vicmap layer as an sf. Filters can be made using a bounding box or with additional CQL filtering statements.

read_layer_sf(
  layer_name,
  filter = NULL,
  boundbox = NULL,
  CRS = 4283,
  Client = newClient(),
  ...
)

Arguments

layer_name
character; name of Vicmap layer/feature to read in.

filter
character; optional CQL filter statement. See CQL tutorial for more information.

boundbox
character; optional string of coordinates to restrict query within. Can be generated using the boundbox function or formatted as a character string following: "xmin,ymin,xmax,ymax".

CRS
numeric; coordinate reference system for query (default is 4283).

Client
Object of R6Class; Vicmap client generated with newClient. Mandatory when using a boundbox; however default is newClient

...
Additional arguments passed to read_sf (e.g. quiet and as_tibble).

Value

sf/data.frame/tibble

Examples

VicmapClient <- newClient()
data <- read_layer_sf(layer_name = "datavic:VMHYDRO_WATERCOURSE_DRAIN",
                      boundbox = boundbox(xmin = 144.25,
                                          ymin = -38.44,
                                          xmax = 144.50,
                                          ymax = -38.25),
                      filter = "HIERARCHY = 'L'",
                      Client = VicmapClient)