options

#' This function retrieves bcdata specific options that can be set. These options can be set using option({name of the option} = {value of the option}). The default options are purposefully set conservatively to hopefully ensure successful requests. Resetting these options may result in failed calls to the data catalogue. Options in R are reset every time R is re-started.

vicmap.max_geom_pred_size is the maximum size of an object used for a geometric operation. Objects that are bigger than this value will be simplified in the request call using sf::st_simplify(). This is done to reduce the size of the query being sent to the WFS geoserver.

vicmap.chunk_limit is an option useful when dealing with very large data sets. When requesting large objects from the catalogue, the request is broken up into smaller chunks which are then recombined after they've been downloaded. VicmapR does this all for you but using this option you can set the size of the chunk requested. On faster internet connections, a bigger chunk limit could be useful while on slower connections, it is advisable to lower the chunk limit. Chunks must be less than 70000.

vicmap.base_url is the base wfs url used to query the geoserver.

vicmap_options()

check_chunk_limit()

Value

vicmap_options() returns a data.frame

Examples

vicmap_options()

#> # A tibble: 3 x 3
#>   option             value    default                                           
#>   <chr>              <list>   <chr>                                             
#> 1 vicmap.max_geom_p… <lgl [1… 4400                                              
#> 2 vicmap.chunk_limit <lgl [1… 70000                                             
#> 3 vicmap.base_url    <NULL>   http://services.land.vic.gov.au/catalogue/publicp…