vicmap_query

Begin a Vicmap WFS query by selecting a WFS layer. The record must be available as a Web Feature Service (WFS) layer (listed in listLayers())

vicmap_query(layer, CRS = 4283, wfs_version = "2.0.0")

Arguments

layer
vicmap layer to query. Options are listed in listLayers()

CRS
Coordinate Reference System (default is 4283)

wfs_version
The current version of WFS is 2.0.0. GeoServer supports versions 2.0.0, 1.1.0, and 1.0.0. However in order for filtering to be correctly applied wfs_version must be 2.0.0 (default is 2.0.0)

Value

vicmap_promise

Details

The returned vicmap_promise object is not data, rather it is a 'promise' of the data that can be returned if collect() is used; which returns an sf object.

Examples

vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN")

#> * Using collect() on this object will return 191460
#> * features and 16 fields
#> * At most six rows of the record are printed here
#> --------------------------------------------------------------------------------
#> Simple feature collection with 6 features and 15 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: 142.7675 ymin: -35.06905 xmax: 143.324 ymax: -35.04559
#> CRS:            4283
#> # A tibble: 6 x 16
#>   id                  PFI    UFI FEATURE_TYPE_CODE NAME  NAMED_FEATURE_ID ORIGIN
#>   <chr>             <int>  <int> <chr>             <chr> <chr>            <chr> 
#> 1 VMHYDRO_WATERC~ 8553127 2.55e6 watercourse_chan~ NA    NA               2     
#> 2 VMHYDRO_WATERC~ 8553130 2.55e6 watercourse_chan~ NA    NA               2     
#> 3 VMHYDRO_WATERC~ 8553143 2.55e6 watercourse_chan~ NA    NA               2     
#> 4 VMHYDRO_WATERC~ 8553149 2.55e6 watercourse_chan~ NA    NA               2     
#> 5 VMHYDRO_WATERC~ 8553158 2.55e6 watercourse_chan~ NA    NA               2     
#> 6 VMHYDRO_WATERC~ 8553168 2.55e6 watercourse_chan~ NA    NA               2     
#> # ... with 9 more variables: CONSTRUCTION <chr>, USAGE <chr>, HIERARCHY <chr>,
#> #   FEATURE_QUALITY_ID <int>, CREATE_DATE_PFI <dttm>, SUPERCEDED_PFI <chr>,
#> #   CREATE_DATE_UFI <dttm>, OBJECTID <int>, geometry <LINESTRING [°]>