Filter Data

See dplyr::filter for details.

# S3 method for vicmap_promise
filter(.data, ...)

Arguments

.data

object of class vicmap_promise (likely passed from vicmap_query())

...

filter statements

Value

Object of class vicmap_promise, which is a 'promise' of the data that can be returned if collect() is used

Methods (by class)

vicmap_promise: filter.vicmap_promise

Examples

# \donttest{
try(
vicmap_query(layer = "open-data-platform:hy_watercourse") %>%
 filter(hierarchy == "L", pfi == 8553127)
 )
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteIdentifier’,
#>  target signature ‘wfsConnection#ident’.
#>  "wfsConnection#ANY" would also be valid
#> Note: method with signature ‘DBIConnection#character’ chosen for function ‘dbQuoteString’,
#>  target signature ‘wfsConnection#character’.
#>  "wfsConnection#ANY" would also be valid
#> • Using collect() on this object will return 1 features
#> • and 15 fields
#> • At most six rows of the record are printed here
#> ────────────────────────────────────────────────────────────────────────────────
#> Simple feature collection with 1 feature and 14 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 143.3209 ymin: -35.06905 xmax: 143.324 ymax: -35.06721
#> Geodetic CRS:  GDA94
#> # A tibble: 1 × 15
#>   id       pfi    ufi featu…¹ name  named…² origin usage hiera…³ featu…⁴ const…⁵
#>   <chr>  <int>  <int> <chr>   <chr> <chr>   <chr>  <chr> <chr>     <int> <chr>  
#> 1 hy_w… 8.55e6 2.55e6 waterc… NA    NA      2      3     L          4762 1      
#> # … with 4 more variables: superceded_pfi <chr>, create_date_pfi <date>,
#> #   create_date_ufi <date>, geometry <LINESTRING [°]>, and abbreviated variable
#> #   names ¹​feature_type_code, ²​named_feature_id, ³​hierarchy,
#> #   ⁴​feature_quality_id, ⁵​construction
 # }