Get Column Information

geom_col_name returns a single value for the name of the geometry column for the WFS layer selected in the vicmap_promise object (e.g. SHAPE). This column will become the geometry column when using collect(). feature_cols() provides a vector of all column names for the WFS layer selected in the vicmap_promise object and get_col_df() returns a data.frame with the column names and their XML schema string datatypes.

geom_col_name(x)

feature_cols(x)

get_col_df(x)

Arguments

x

object of class vicmap_promise

Value

character/data.frame

Examples

# \donttest{
# Return the name of the geometry column
vicmap_query(layer = "open-data-platform:hy_watercourse") %>% 
  geom_col_name()
#> [1] "geom"
 # }
# \donttest{
# Return the column names as a character vector
vicmap_query(layer = "open-data-platform:hy_watercourse") %>% 
  feature_cols()
#>  [1] "pfi"                "ufi"                "feature_type_code" 
#>  [4] "named_feature_id"   "origin"             "construction"      
#>  [7] "usage"              "hierarchy"          "feature_quality_id"
#> [10] "create_date_pfi"    "superceded_pfi"     "create_date_ufi"   
#> [13] "geom"              
# }   
# \donttest{
# Return a data.frame of the columns and their XML schema string datatypes
try(
vicmap_query(layer = "open-data-platform:hy_watercourse") %>% 
  get_col_df()
  )
#>                  name                  type
#> 1                 pfi           xsd:decimal
#> 2                 ufi           xsd:decimal
#> 3   feature_type_code            xsd:string
#> 4    named_feature_id           xsd:decimal
#> 5              origin            xsd:string
#> 6        construction            xsd:string
#> 7               usage            xsd:string
#> 8           hierarchy            xsd:string
#> 9  feature_quality_id           xsd:decimal
#> 10    create_date_pfi              xsd:date
#> 11     superceded_pfi           xsd:decimal
#> 12    create_date_ufi              xsd:date
#> 13               geom gml:CurvePropertyType
 # }