Show The Query

show_query() summarises the constructed query that has been passed to it by printing details about the query in a human readable format.

See dplyr::show_query for details.

# S3 method for vicmap_promise
show_query(x, ...)

Arguments

x
Object of class vicmap_promise (likely passed from vicmap_query())

...
Other parameters possibly used by generic

Value

object of class vicmap_promise (invisible: query printed to console), which is a 'promise' of the data that can be returned if collect() is used

Details

The printed information consists of three sections:

base url The base url of the query, this can be changed with options(vicmap.base_url = another_url)

body Lists the parameters of the WFS query, these can be modified through various functions such as vicmap_query(), filter(), select() and head()

full query url The constructed url of the final query to be collected

Methods (by class)

vicmap_promise: show_query.vicmap_promise

Examples

# \donttest{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>%
head(50) %>%
show_query()

#> <base url>
#> http://services.land.vic.gov.au/catalogue/publicproxy/guest/dv_geoserver/wfs
#> 
#> <body>
#> service: wfs 
#> version: 2.0.0 
#> request: GetFeature 
#> typeNames: datavic:VMHYDRO_WATERCOURSE_DRAIN 
#> outputFormat: application/json 
#> count: 50 
#> srsName: EPSG:4283 
#> 
#> <full query url>
#> http://services.land.vic.gov.au/catalogue/publicproxy/guest/dv_geoserver/wfs?service=wfs&version=2.0.0&request=GetFeature&typeNames=datavic%3AVMHYDRO_WATERCOURSE_DRAIN&outputFormat=application%2Fjson&count=50&srsName=EPSG%3A4283

# }